Delete .gitignored file .env
On my machine, changes to .env are visible if I run `git status`. I
think we should delete all .gitignored files and clearly state in the
README what needs to be done.

Also the .env.template did not have a JWT secret. Probably because sb.
edited the local .env file and forgot to edit it in the template.