Configure Travis CI conditional builds
and add placeholder deployment. This commit should deploy to

* `production` if on branch master AND tagged
* `staging` if on branch master
* `custom` if tagged

The custom deployment is inspired by https://zeit.co/now and should
deploy features for demonstration purposes before merging into master.

It will happen that we trigger multiple deployments with this workflow:
If we deploy to production, that is going to deploy to a custom
environment and to staging. But I think this is not too problematic.

What do you think @appinteractive @mcwimpy?