Remove remains of cypress
I love cypress already. However my suggestion is to create a separate
repository that includes `backend` and `web` run full stack tests from
there. It is possible to trigger builds in other repositories on Travis
CI. So here is my suggestion: We agree to have matching branch names in
all related repositories, so `branch-1` in repo `web`, `backend` and
`human-connection`.

As a start, the developer has to manually update the submodules in the
meta repository and trigger a build to run fullstack tests. In the
future we might automate this process by triggering builds in other
repos. That's possible on Travis CI, see:
https://docs.travis-ci.com/user/triggering-builds/

See each repository is responsible to keep their set of unit/integration
tests but the meta repository is responsible of carrying out full stack
testing.

@appinteractive your opinion?