Split development config from production config
in `docker-compose.yml` and `docker-compose.override.yml`. This should speed
up builds e.g. on Travis CI, which does not need to sync folders or run
`yarn run dev` if the docker image was built recently. Also it should
make the build more reliable as it behaves more similar to our deployment.