Refactor `login` with confidence
This is why I :heart: testing. First I wrote several tests to get
familiar with the existing code and specified everything that I believe to
be required. Then I refactor the entire method at once and the tests pass.

Note that I do **not** believe that setting the user data to `null` is
really relevant.