Fix non-deterministic erros in testing
Without the `--runInBand` flag, jest will create a worker pool of child
processes and run the tests withing. No surprise, why the tests were
running completely non-deterministically! If we clear the database in
the middle of the tests we will end up with no records, or if we run the
factories in several tests we end up with many duplicate records.

FYI: @mattwr18 @appinteractive @tansaku @Kachulio1