CMD and EXPOSE is valid for all docker stages
FYI: @mattwr18 and @ksinkar
I wanted to test cypress with the Travis configuration. I had
to add `command: ...` to the `docker-compose.travis.yml` which I find
quite annoying. So I put CMD and EXPOSE at the top of the Dockerfile,
thus making it a default for all stages including build-and-test.