Add commits to configuration to deploy :latest tag
The recommended way to update a kubernetes deployment to a new image is
to change the image tag. However, our build server is configured to push
every commit of the `master` branch to docker hub to the respective
repository and tag `:latest`. So adding some configuration that can be
changed seems to be a trick to re-deploy the `:latest` image.

See here:
https://stackoverflow.com/a/51835397