Refactoring: No uploads/ vol,access neo4j directly
Digital Ocean does not support shared directories. So we have to uploads
the images in `/uploads` via `kubectl cp` or something similar.
Likewise, it is not possible to share the exported mongodb .json files
with neo4j container. Therefore let's install `cypher-shell`, included
in `neo4j` package, to directly open a neo4j connection and bulk import
the data.