Small changes on translation mistakes - Portuguese
The translation of "post" is "postagem", and not "pastagem". Pastagem means where the cow eats grass.
The "admin" was as "Administrator", while the correct word is "Administrador".

I'm not sure about what "Shouts" are, but I don't really know if the word "Gritos" is the properly for this thing. In portuguese, when you say "Grito", the literal translation of "Shouts" you think of a person screaming, and I don't think that's was the intention. If somebody could explain me what a "Shout" means in this code I could help in a more accurate translation.