Trigger rollout with a change to spec.template
The documentation clearly says:
```
Note: A Deployment’s rollout is triggered if and only if the
Deployment’s pod template (that is, .spec.template) is changed, for
example if the labels or container images of the template are updated.
Other updates, such as scaling the Deployment, do not trigger a
rollout.
```

Read: https://kubernetes.io/docs/concepts/workloads/controllers/deployment/#updating-a-deployment