Add CLIENT_URI to the to-be-tested server
Our ActivityPub service needs to know the public URI how 3rd party
servers can reach the server.

At the moment, we have this configuration:
https://nitro-staging.human-connection.org/activityPub (external)
=> http://nitro-backend.human-connection:4000 (internal)

So, when the backend generates links, they will have the external URI as
a prefix.

This commit might be a little dirty, because CLIENT_URI will be
different now for frontend and backend during testing. I hope this does
not add extra confusion. It seems that our cypress tests are not
affected by this change.