Disabling activityPub middleware again
I thought it might be a good thing to create RSA keys at least for
cucumber testing. At least that's where we test activitityPub. But it
caused too flaky tests and broken builds. @Mastercuber if you need to
create a user and you need the RSA keys, don't create it through seeds
but through the actual server.