Fix test, remove vue-filter `relativeDateTime`
cc @Tirokk @ulfgebhardt @appinteractive

This is a perfect example why I insist on writing tests. While debugging
@Tirokk and I discovered several design flaws and plugins that depend on
each other. The solution to all of this is not to use vue-filters plugin
at all! Vue-filters depends on Vuex, i18n, vuex-i18n, nuxt-modules and so
on.

This is just bad, bad, bad code. Start writing tests. Now.

We should start to refactor vue-filters and use components instead.