Silence expected error messages on build server
@ulfgebhardt we have a quite lengthy log output on our build server.
This `2> /dev/null` prevents the server from writing log output for
expected errors. E.g. we're testing if a user can log in with incorrect
credentials or we try to create duplicates for unique slugs. All of this
will output errors to the command line which we're not interested in.

We could think of configuring the `package.json` in a way that outputs
these error messages in case we set an environment variable:
```
DEBUG=1 yarn run test:jest
```
That would be an option. Or use the `debug` npm packages that
@Mastercuber has introduced.