Revert logout.vue
For any static image URL we don't need to use `<hc-image>`. The assets
will be served from the frontend container. I did not anticipate that we
have a lot of static images in the frontend when I said "You can remove
every image with the `<hc-image>` component".

@aonomike: In the future you can make some screenshots and post them in the PR
and ask for help.

@ulfgebhardt: Yes, it would have been safer not to merge the PR. We
might run into the danger of forgetting things like this.