Add thumbor service
This is my birthday present to @mattwr18. Feel free to merge this in
your PR unless you think this extends the scope of the PR.

Yet, I don't know how to send the returned URL of the image (it's in the
response header) to the backend.

Either this
1. webapp <---> thumbor
2. webapp <---> backend

Or this
1. webapp <---> backend
2. backend <---> thumbor