Move the filter menu into the index page
A filter makes sense only on the news feed page in my opinion. As a
comparison: The way how full text search is implemented makes sense in
the menu bar, which is always visible. Why? The search results are shown
in a popover. Therefore those are different from the news feed. The news
feed however is only visible on the index page. Therefore I believe, it
makes sense to put the filter menu there and only there.