Fix bug introduced in 6f0447515
@ulfgebhardt For the `deleted` and `disabled` flags I would be glad if we could
take only one value, that is `true`, in case the post is considered disabled or
deleted respectively.

There might be more bugs like this now. Please keep in mind that these
counts are also used as total limits for pagination and you break the
pagination if you change the cypher statement.