Page size of 12 entries on / page
It annoys me that there are gaps for the 3 column layout, even if there
are more records left. So, the newsfeed layout has 1 column on mobile, 2
columns on tablet, and 3 columns on a desktop pc. So we should choose a
page size which is dividable by 1,2 and 3 that is 6. I have chosen 12
which is 2*6 to avoid clicking the reload posts button so often.