Fix @Tirokk's suggestions
Wow @Tirokk that was a good catch! I thought that the emitting side
knows the `index`, but no, it only emits `removePostFromList` and that's
it. In order to delete the right post, we have to pass in the `index`
ourselves. So, the caller of the function is the page component itself.

I didn't see the error myself, because in Javascript if you parse
`undefined` or `null` as Integer, you get 0 and it does not raise an
error.