Fix #1394
This bug is actually an unpleasant combination of bugs. We ran into:
https://github.com/Akryum/vue-apollo/issues/492

The people there suggest to use `no-cache`. But that would let us face
yet another bug:
https://github.com/Akryum/vue-apollo/issues/263

So I fixed this with an obscure workaround and assumed that whenever the
two results intersect then this is a cache update.