Add backend test for @mattwr18
This was already covered by `slugifyMiddleware.spec.js`. I do agree that
this is a better place for it. Also we're not testing GraphQL requests
here but the factories. As we're testing the existence of unique
constraints, I think it won't matter if we use factories or actual
GraphQL requests.