build: less verbose output on build server
I'm annoyed to scroll when I'm debugging the build server logs. I think,
if you run `jest` with an environment variable `CI` set to true, it will
not try to output colorized log and also does not output which test it's
currently running.

That should lead to the desired, smaller log on Travis CI.