Decrease width of language select, pull right
- @ogerly hope you are cool with this, I pulled it right so that the
validations you added still all line up on the right side

- @alina-beck I think you know a more sophisticated way to do this, it
was just easy for me to add an empty <ds-flex-item> with a 50% width on
medium and larger devices. This can also be populated with country of
contribution or visibility, when one of those is implemented.. what do
you think?