Show the comments again
@alina-beck the CSS class `.ds-card-footer` gets in the way of the
`overflow: hidden` from the styleguide. The scoping avoids the issue.

Of course this is not a great solution and the whole file needs to be
refactored and fixed. I also see a lot of `<div>`s with a style
attribute. All of that is legacy code that survived for too long.