fix #2229
I have to agree with @vbelolapotkov here. First of all the translation
for German an English does not match and second it does not make an awful
lot of sense to have "Report these actions" twice in the list. The
English translation "Advocacy or encouragement to these behaviors."
probably belongs into the section before.