refactor(translations): Fix keys for muting
Provide the correct German translation. Remove "" and use `null` for
Italian. @mattwr18 we really have to make sure not to add empty strings
to our translations because we disable the fallback in that case. Also,
if we want, we could replace the other translations with `null` in order
to make sure that we have the better (though untranslated) explanation.

@mattwr18 I had a look into `notificationsMiddleware.js`. I think that
we can still keep the tests and the behaviour although there is no
`BLOCKED` relationship anymore, right?