Replace buildList with array of Promises
reason: Rosie.js does not support promises. So for the time being, we
shall not use more than one `after` callback and we shall not use
buildList as it does not return a `Promise.all([])`. We have to ask the
maintainers of Rosie.js if they accept the breaking change of returning
a Promise always.